% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tomorrowio_timeline}
\alias{tomorrowio_timeline}
\title{Call Tomorrow.io Timeline API}
\usage{
tomorrowio_timeline(
  api_key,
  location,
  fields,
  timesteps,
  start_time,
  end_time,
  units = "metric",
  timezone = "UTC",
  batch_size_days = 0
)
}
\arguments{
\item{api_key}{`string` API key use to access [Tomorrow.io](https://www.tomorrow.io/weather-api/) services.}

\item{location}{2 element `vector` of latitude and longitude location}

\item{fields}{`list` of `string`s of [field names](https://docs.tomorrow.io/reference/data-layers-overview)}

\item{timesteps}{`string` [time step](https://docs.tomorrow.io/reference/data-layers-overview#timestep-availability)}

\item{start_time, end_time}{`POSIXct` start and end date and time inclusive to get data for.}

\item{units}{`string` unit name, possible values are `metric` or `imperial` (default is `metric`)}

\item{timezone}{`string` timezone in ISO 8601 format (default is `"UTC"`).}

\item{batch_size_days}{`numeric` batch size in days. When fetching a large data set, it is often
better to make several API calls and combine the results. Passing in a value other than `0`
will split into fixed sized batchs of the specified size in days. `0` will create a single
batch (default is `0` which will make a single call to Timeline API, if needed, and multiple 29
day calls to Historical, if needed).}
}
\value{
A `data.frame`.
}
\description{
Fetch current and forecast weather data.
}
\examples{
tomorrowio_timeline(
  api_key = API_KEY, location = c(42.3478, -71.0466), fields = c("temperature"), timesteps = c("1d"),
  start_time = Sys.time() - (60 * 60 * 24 * 3), end_time = Sys.time() + (60 * 60 * 24 * 5)
)
}
